+++ api/database/models/Review.js
@@ -20,9 +20,9 @@ module.exports = (sequelize, DataTypes) => {
     paranoid: true,
     timestamps: true,
     modelName: 'Review',
-  },Review.addHook('beforeSave', async (review) => {
-    return review.id = uuid();
-  })
-  );
+  });
+  Review.addHook('beforeSave', async (review) => {
+      return review.id = uuid();
+    })
   return Review;
 };
