+++ package.json
@@ -1,0 +1,5 @@
+{
+  "devDependencies": {
+    "prettier": "2.7.1"
+  }
+}
